separate logic into a utility function for logout and check if user is authenticated
