+++ src/utils/auth/token.js
@@ -1,0 +1,10 @@
+const get = () => localStorage.getItem('OS_AUTH_TOKEN');
+
+const remove = () => {
+  localStorage.removeItem('OS_AUTH_TOKEN');
+};
+
+export default {
+  get,
+  remove,
+};
